+++ src/web/component/BonusNumberInput/BonusNumberInput.js
@@ -1,6 +1,7 @@
-import { $ } from '../util/domSelector';
-import numericInputFilter from '../util/numericInputFilter';
-import { SETTING } from '../../constant/setting';
+import { $ } from '../../util/domSelector';
+import numericInputFilter from '../../util/numericInputFilter';
+import { SETTING } from '../../../constant/setting';
+import styles from './BonusNumberInput.module.css';
 
 class BonusNumberInput extends HTMLElement {
   connectedCallback() {
@@ -9,7 +10,8 @@ class BonusNumberInput extends HTMLElement {
   }
 
   #setNumericInputFilter() {
-    $('.input-group', this).addEventListener('input', (event) =>
+    const inputGroupClass = `.${styles['input-group']}`;
+    $(inputGroupClass, this).addEventListener('input', (event) =>
       numericInputFilter(event.target, SETTING.MAX_LOTTO_NUMBER_LENGTH),
     );
   }
@@ -17,8 +19,8 @@ class BonusNumberInput extends HTMLElement {
   #render() {
     this.innerHTML = `
       <div>
-        <label>보너스 번호</label>
-        <div class="input-group flow-right">
+        <label class="${styles.label}">보너스 번호</label>
+        <div class="${styles['input-group']}">
           <input class="bonus-number-input" type="number" \
           min=${SETTING.MIN_LOTTO_NUMBER} max=${SETTING.MAX_LOTTO_NUMBER} />
         </div>
