+++ src/web/component/BonusNumberInput/BonusNumberInput.module.css
@@ -1,0 +1,17 @@
+.label {
+  display: block;
+  margin: 0 0 6px;
+}
+
+.input-group {
+  display: flex;
+  flex-flow: row wrap;
+  gap: 10px;
+  flex-direction: row-reverse;
+}
+
+.input-group input {
+  width: 34px;
+  font-size: var(--font-caption-size);
+  text-align: center;
+}
