+++ src/web/component/LottoGameApp/LottoGameApp.module.css
@@ -1,0 +1,29 @@
+lotto-game-app {
+  background: var(--color-greyscale-2);
+  height: 100%;
+  display: flex;
+  flex-flow: column nowrap;
+  padding: 30px 0;
+  align-items: center;
+}
+
+.lotto-game {
+  background: var(--color-greyscale-1);
+  margin: auto;
+  border: 1px solid var(--color-lotto-game-border);
+  border-radius: 4px;
+  padding: 16px;
+  width: 414px;
+  height: 727px;
+  display: flex;
+  flex-flow: column nowrap;
+  gap: 28px;
+  overflow-y: auto;
+}
+
+.lotto-game h1 {
+  color: var(--color-greyscale-5);
+  text-align: center;
+  margin-top: 28px;
+}
+
