+++ src/web/component/PurchaseForm/PurchaseForm.js
@@ -1,8 +1,9 @@
-import { $ } from '../util/domSelector';
-import ErrorMessageUtil from '../util/ErrorMessageUtil';
-import numericInputFilter from '../util/numericInputFilter';
-import Validator from '../../validator/Validator';
-import { SETTING } from '../../constant/setting';
+import { $ } from '../../util/domSelector';
+import ErrorMessageUtil from '../../util/ErrorMessageUtil';
+import numericInputFilter from '../../util/numericInputFilter';
+import Validator from '../../../validator/Validator';
+import { SETTING } from '../../../constant/setting';
+import styles from './PurchaseForm.module.css';
 
 class PurchaseForm extends HTMLElement {
   #elements;
@@ -61,9 +62,9 @@ class PurchaseForm extends HTMLElement {
 
   #render() {
     this.innerHTML = `
-      <section id="purchase">
+      <section class="${styles.purchase}">
         <p>구입할 금액을 입력해주세요.</p>
-        <form id="purchase-form">
+        <form id="purchase-form" class="${styles['purchase-form']}">
           <input id="purchase-form-input" type="number" placeholder="금액">
           <button type="button" id="purchase-form-button">구입</button>
         </form>
