+++ src/web/component/WinningNumbersInput/WinningNumbersInput.js
@@ -1,6 +1,7 @@
-import { $ } from '../util/domSelector';
-import numericInputFilter from '../util/numericInputFilter';
-import { SETTING } from '../../constant/setting';
+import { $ } from '../../util/domSelector';
+import numericInputFilter from '../../util/numericInputFilter';
+import { SETTING } from '../../../constant/setting';
+import styles from './WinningNumbersInput.module.css';
 
 class WinningNumbersInput extends HTMLElement {
   connectedCallback() {
@@ -9,7 +10,8 @@ class WinningNumbersInput extends HTMLElement {
   }
 
   #setNumericInputFilter() {
-    $('.input-group', this).addEventListener('input', (event) =>
+    const inputGroupClass = `.${styles['input-group']}`;
+    $(inputGroupClass, this).addEventListener('input', (event) =>
       numericInputFilter(event.target, SETTING.MAX_LOTTO_NUMBER_LENGTH),
     );
   }
@@ -25,8 +27,8 @@ class WinningNumbersInput extends HTMLElement {
   #render() {
     this.innerHTML = `
       <div>
-        <label>당첨 번호</label>
-        <div class="input-group">
+        <label class="${styles.label}">당첨 번호</label>
+        <div class="${styles['input-group']}">
           ${this.#createInputGroup().join('')}
         </div>
       </div>
