--- src/web/css/style.css
@@ -1,330 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-  border: 0;
-  font: inherit;
-  color: inherit;
-  -webkit-box-sizing: border-box;
-  -moz-box-sizing: border-box;
-  box-sizing: border-box;
-}
-
-input::-webkit-outer-spin-button,
-input::-webkit-inner-spin-button {
-  -webkit-appearance: none;
-  margin: 0;
-}
-
-input[type="number"] {
-  -moz-appearance: textfield;
-}
-
-:root {
-  --color-primary: rgb(78 91 166);
-  --color-greyscale-1: rgb(255 255 255);
-  --color-greyscale-2: rgb(252 252 253);
-  --color-greyscale-3: rgb(180 180 180);
-  --color-greyscale-4: rgb(139 139 139);
-  --color-greyscale-5: rgb(0 0 0);
-  --color-lotto-game-border: rgba(0 0 0 / 12%);
-  --color-error-text: rgb(255 0 0);
-  --font-title-size: 24px;
-  --font-subtitle-size: 20px;
-  --font-body-size: 15px;
-  --font-placeholder-size: 15px;
-  --font-caption-size: 14px;
-  --font-title-weight: 800;
-  --font-subtitle-weight: 600;
-  --font-caption-weight: 700;
-}
-
-html {
-  font-family: Roboto, sans-serif;
-  height: 100vh;
-}
-
-body {
-  height: calc(100% - (64px + 80px));
-}
-
-lotto-header {
-  background: var(--color-primary);
-  padding: 0 130px;
-  height: 64px;
-  display: flex;
-  align-items: center;
-}
-
-lotto-game-app {
-  background: var(--color-greyscale-2);
-  height: 100%;
-  display: flex;
-  flex-flow: column nowrap;
-  padding: 30px 0;
-  align-items: center;
-}
-
-button {
-  background: var(--color-primary);
-  color: var(--color-greyscale-1);
-  font-size: var(--font-caption-size);
-  font-weight: var(--font-caption-weight);
-  letter-spacing: 1.25px;
-  border-radius: 4px;
-  padding: 4px 6px;
-}
-
-button:hover {
-  cursor: pointer;
-  transition: all .3s ease-in-out;
-  opacity: 0.85;
-}
-
-button:disabled {
-  background: var(--color-greyscale-4);
-  cursor: inherit;
-}
-
-button:disabled:hover {
-  opacity: inherit;
-}
-
-h1 {
-  font-weight: var(--font-title-weight);
-  font-size: var(--font-title-size);
-  color: var(--color-greyscale-1);
-}
-
-h2 {
-  font-size: var(--font-subtitle-size);
-  font-weight: var(--font-subtitle-weight);
-  color: var(--color-greyscale-5);
-  text-align: center;
-}
-
-input {
-  border: 1px solid var(--color-greyscale-3);
-  border-radius: 4px;
-  padding: 8px;
-  height: 36px;
-}
-
-input:disabled {
-  background: rgba(0 0 0 / 5%);
-}
-
-.error-text {
-  font-size: 12px;
-  color: var(--color-error-text);
-  letter-spacing: 0;
-  word-break: keep-all;
-  animation: fade-in 0.5s;
-}
-
-#lotto-game {
-  background: var(--color-greyscale-1);
-  margin: auto;
-  border: 1px solid var(--color-lotto-game-border);
-  border-radius: 4px;
-  padding: 16px;
-  width: 414px;
-  height: 727px;
-  display: flex;
-  flex-flow: column nowrap;
-  gap: 28px;
-  overflow-y: auto;
-}
-
-#lotto-game h1 {
-  color: var(--color-greyscale-5);
-  text-align: center;
-  margin-top: 28px;
-}
-
-#purchase {
-  font-size: var(--font-body-size);
-  line-height: 24px;
-  letter-spacing: 0.5px;
-  animation: fade-in 0.5s;
-}
-
-#purchase form {
-  display: flex;
-  flex-flow: row wrap;
-  align-items: flex-end;
-  gap: 10px;
-}
-
-#purchase input {
-  width: 314px;
-  gap: 10px;
-}
-
-#purchase button {
-  padding: 6px 8px;
-  width: 56px;
-}
-
-#purchase-result {
-  font-size: var(--font-body-size);
-  line-height: 24px;
-  letter-spacing: 0.5px;
-  height: 300px;
-  animation: fade-in 0.5s;
-}
-
-#purchase-result ul {
-  max-height: 276px;
-  overflow: scroll;
-}
-
-#purchase-result li {
-  list-style-type: none;
-  margin: 6px 0;
-}
-
-#purchase-result li::before {
-  content: '🎟️';
-  font-size: 24px;
-  margin-right: 7px;
-  vertical-align: middle;
-}
-
-#winning-numbers {
-  font-size: var(--font-body-size);
-  line-height: 24px;
-  letter-spacing: 0.5px;
-  display: flex;
-  flex-flow: column wrap;
-  gap: 10px;
-  animation: bottom-up 0.5s;
-}
-
-#winning-numbers form {
-  display: flex;
-  flex-flow: row wrap;
-  justify-content: space-between;
-  gap: 6px;
-}
-
-#winning-numbers label {
-  display: block;
-  margin: 0 0 6px;
-}
-
-#winning-numbers input {
-  width: 34px;
-  font-size: var(--font-caption-size);
-  text-align: center;
-}
-
-#winning-numbers .input-group {
-  display: flex;
-  flex-flow: row wrap;
-  gap: 10px;
-}
-
-#winning-numbers .input-group.flow-right {
-  flex-direction: row-reverse;
-}
-
-#winning-numbers #bonus-number-input-group {
-  text-align: right;
-}
-
-#winning-numbers button {
-  width: 100%;
-  margin-top: 15px;
-}
-
-dialog {
-  background: var(--color-greyscale-1);
-  width: 350px;
-  height: 500px;
-  margin: auto;
-  border-radius: 4px;
-  animation: bottom-up 0.7s;
-}
-
-dialog::backdrop {
-  background: rgba(0 0 0 / 50%);
-  animation: fade-in 0.7s;
-}
-
-#result-modal-box {
-  width: 100%;
-  height: 100%;
-  padding: 40px 16px 32px;
-  display: flex;
-  flex-flow: column nowrap;
-  overflow: scroll;
-  gap: 32px;
-  text-align: center;
-}
-
-#result-modal-box table {
-  border-collapse: collapse;
-}
-
-#result-modal-box thead {
-  font-weight: var(--font-subtitle-weight);
-}
-
-#result-modal-box tr {
-  border-top: 1px solid var(--color-lotto-game-border);
-  border-bottom: 1px solid var(--color-lotto-game-border);
-}
-
-#result-modal-box th,
-#result-modal-box td {
-  padding: 8px;
-  height: 41px;
-}
-
-#result-modal-box #close-modal-button {
-  position: absolute;
-  top: 16px;
-  right: 16px;
-  height: 14px;
-  cursor: pointer;
-}
-
-#result-modal-box #profit-rate-text {
-  font-weight: var(--font-caption-weight);
-}
-
-#result-modal-box #restart-button {
-  padding: 10px;
-}
-
-footer {
-  border-top: 1px solid #4E5BA633;
-  height: 80px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-footer p {
-  font-weight: var(--font-caption-weight);
-  font-size: var(--font-caption-size);
-  color: var(--color-primary);
-  letter-spacing: 1.25px;
-}
-
-@keyframes fade-in {
-  from { opacity: 0; }
-  to { opacity: 1; }
-}
-
-@keyframes bottom-up {
-  from {
-    bottom: -20px;
-    opacity: 0;
-  }
-
-  to {
-    bottom: 0;
-    opacity: 1;
-  }
-}
